requirements.txt
einops==0.6.1
fairscale==0.4.13
filelock==3.13.1
ftfy==6.1.3
h5py==3.10.0
imgaug==0.2.6
keras==2.11.0
kornia==0.7.2
kornia_rs==0.1.2
lmdb==1.4.1
matplotlib==3.7.4
matplotlib-inline==0.1.6
numpy==1.24.3
omegaconf==2.3.0
open-clip-torch==2.24.0
openai-clip==1.0.1
openpyxl==3.1.2
pandas==2.0.3
Pillow==9.5.0
safetensors==0.4.1
scikit-image==0.20.0
scikit-learn==1.3.2
scipy==1.9.1
sentencepiece==0.2.0
streamlit==1.30.0
tenacity==8.2.3
tensorboard==2.11.2
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorboardX==2.6.2.2
timm==0.9.6
torch>=2.0.1
torch-fidelity==0.3.0
torchmetrics==0.6.0
torchsummary==1.5.1
torchvision>=0.15.2
tqdm==4.66.1
